requirements.txt
streamlit
python-dotenv
langchain-core
langchain-mistralai
tavily-python
pydantic
mistralai